+++ src/editor/less/editor.less
@@ -6,6 +6,8 @@
 
 .loadAddonVariables();
 
+// TODO: move these to less variables
+
 & {
   .slate-inline-toolbar {
     position: absolute;
